+++ test/Builder/SimpleRefundBuilder.ts
@@ -1,72 +1,45 @@
-import { MockVaultManager } from '../typechain-types';
-import { DealProvider } from '../typechain-types';
-import { LockDealNFT } from '../typechain-types';
-import { LockDealProvider } from '../typechain-types';
-import { TimedDealProvider } from '../typechain-types';
-import { SimpleBuilder } from '../typechain-types';
-import { deployed, token } from './helper';
+import { MockVaultManager } from '../../typechain-types';
+import { DealProvider } from '../../typechain-types';
+import { LockDealNFT } from '../../typechain-types';
+import { LockDealProvider } from '../../typechain-types';
+import { TimedDealProvider } from '../../typechain-types';
+import { CollateralProvider } from '../../typechain-types';
+import { RefundProvider } from '../../typechain-types';
+import { SimpleRefundBuilder } from '../../typechain-types';
+import { deployed, token, BUSD } from '.././helper';
+import { _createUsers } from './BuilderHelper';
 import { time } from '@nomicfoundation/hardhat-network-helpers';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import { expect } from 'chai';
 import { BigNumber } from 'ethers';
 import { ethers } from 'hardhat';
 
-describe('Simple Builder tests', function () {
+describe('Simple Refund Builder tests', function () {
   let lockProvider: LockDealProvider;
   let dealProvider: DealProvider;
   let mockVaultManager: MockVaultManager;
   let timedProvider: TimedDealProvider;
-  let simpleBuilder: SimpleBuilder;
+  let simpleRefundBuilder: SimpleRefundBuilder;
   let lockDealNFT: LockDealNFT;
-  let userData: SimpleBuilder.BuilderStruct;
+  let userData: SimpleRefundBuilder.BuilderStruct;
   let addressParams: [string, string];
   let projectOwner: SignerWithAddress;
   let startTime: BigNumber, finishTime: BigNumber;
   const amount = ethers.utils.parseEther('100').toString();
   const ONE_DAY = 86400;
-
-  // helper functions
-  async function _createUsers(amount: string, userCount: string): Promise<SimpleBuilder.BuilderStruct> {
-    const pools = [];
-    const length = parseInt(userCount);
-    // Create signers
-    for (let i = 0; i < length; ++i) {
-      const privateKey = ethers.Wallet.createRandom().privateKey;
-      const signer = new ethers.Wallet(privateKey);
-      const user = signer.address;
-      pools.push({ user: user, amount: amount });
-    }
-    const totalAmount = ethers.BigNumber.from(amount).mul(length);
-    return { userPools: pools, totalAmount: totalAmount };
-  }
+  let refundProvider: RefundProvider;
+  let collateralProvider: CollateralProvider;
+  let vaultId: number;
+  let poolId: number;
 
   async function _testMassPoolsData(provider: string, amount: string, userCount: string, params: BigNumber[]) {
     userData = await _createUsers(amount, userCount);
     const lastPoolId = (await lockDealNFT.totalSupply()).toNumber();
-    await simpleBuilder.connect(projectOwner).buildMassPools(addressParams, userData, params);
-    await _logGasPrice(params);
     let k = 0;
     params.splice(0, 0, ethers.BigNumber.from(amount));
     if (provider == timedProvider.address) {
       params.push(ethers.BigNumber.from(amount));
     }
-    for (let i = lastPoolId; i < userData.userPools.length + lastPoolId; i++) {
-      const data = await lockDealNFT.getData(i);
-      expect(data.provider).to.equal(provider);
-      expect(data.poolId).to.equal(i);
-      expect(data.owner).to.equal(userData.userPools[k++].user);
-      expect(data.token).to.equal(token);
-      expect(data.params).to.deep.equal(params);
-    }
-  }
-
-  async function _logGasPrice(params: BigNumber[]) {
-    const tx = await simpleBuilder.connect(projectOwner).buildMassPools(addressParams, userData, params);
-    const txReceipt = await tx.wait();
-    const gasUsed = txReceipt.gasUsed;
-    const GREEN_TEXT = '\x1b[32m';
-    console.log(`${GREEN_TEXT}Gas Used: ${gasUsed.toString()}`);
-    console.log(`Price per one pool: ${gasUsed.div(userData.userPools.length)}`);
   }
 
   function _createProviderParams(provider: string) {
@@ -86,12 +59,21 @@ describe('Simple Builder tests', function () {
     dealProvider = await deployed('DealProvider', lockDealNFT.address);
     lockProvider = await deployed('LockDealProvider', lockDealNFT.address, dealProvider.address);
     timedProvider = await deployed('TimedDealProvider', lockDealNFT.address, lockProvider.address);
-    simpleBuilder = await deployed('SimpleBuilder', lockDealNFT.address);
+    collateralProvider = await deployed('CollateralProvider', lockDealNFT.address, dealProvider.address);
+    refundProvider = await deployed('RefundProvider', lockDealNFT.address, collateralProvider.address);
+    simpleRefundBuilder = await deployed(
+      'SimpleRefundBuilder',
+      lockDealNFT.address,
+      refundProvider.address,
+      collateralProvider.address,
+    );
+    await lockDealNFT.setApprovedProvider(refundProvider.address, true);
     await lockDealNFT.setApprovedProvider(lockProvider.address, true);
     await lockDealNFT.setApprovedProvider(dealProvider.address, true);
     await lockDealNFT.setApprovedProvider(timedProvider.address, true);
+    await lockDealNFT.setApprovedProvider(collateralProvider.address, true);
     await lockDealNFT.setApprovedProvider(lockDealNFT.address, true);
-    await lockDealNFT.setApprovedProvider(simpleBuilder.address, true);
+    await lockDealNFT.setApprovedProvider(simpleRefundBuilder.address, true);
   });
 
   beforeEach(async () => {
@@ -101,55 +83,55 @@ describe('Simple Builder tests', function () {
     finishTime = startTime.add(7 * ONE_DAY); // plus 7 days from `startTime`
   });
 
-  it('should create 10 dealProvider pools', async () => {
+  it('should create 10 simple refund pools with dealProvider', async () => {
     const userCount = '10';
     const params = _createProviderParams(dealProvider.address);
     await _testMassPoolsData(dealProvider.address, amount, userCount, params);
   });
 
-  it('should create 50 dealProvider pools', async () => {
+  it('should create 50 simple refund pools with dealProvider', async () => {
     const userCount = '50';
     const params = _createProviderParams(dealProvider.address);
     await _testMassPoolsData(dealProvider.address, amount, userCount, params);
   });
 
-  it('should create 100 dealProvider pools', async () => {
+  it('should create 100 simple refund pools with dealProvider', async () => {
     const userCount = '100';
     const params = _createProviderParams(dealProvider.address);
     await _testMassPoolsData(dealProvider.address, amount, userCount, params);
   });
 
-  it('should create 10 lockProvider pools', async () => {
+  it('should create 10 simple refund pools with lockProvider', async () => {
     const userCount = '10';
     const params = _createProviderParams(lockProvider.address);
     await _testMassPoolsData(lockProvider.address, amount, userCount, params);
   });
 
-  it('should create 50 lockProvider pools', async () => {
+  it('should create 50 simple refund pools with lockProvider', async () => {
     const userCount = '50';
     const params = _createProviderParams(lockProvider.address);
     await _testMassPoolsData(lockProvider.address, amount, userCount, params);
   });
 
-  it('should create 100 lockProvider pools', async () => {
+  it('should create 100 simple refund pools with lockProvider', async () => {
     const userCount = '100';
     const params = _createProviderParams(lockProvider.address);
     await _testMassPoolsData(lockProvider.address, amount, userCount, params);
   });
 
-  it('should create 10 timedProvider pools', async () => {
+  it('should create 10 simple refund pools with timedProvider', async () => {
     const userCount = '10';
     const params = _createProviderParams(timedProvider.address);
     await _testMassPoolsData(timedProvider.address, amount, userCount, params);
   });
 
-  it('should create 50 timedProvider pools', async () => {
+  it('should create 50 simple refund pools with timedProvider', async () => {
     const userCount = '50';
     const params = _createProviderParams(timedProvider.address);
     await _testMassPoolsData(timedProvider.address, amount, userCount, params);
   });
 
-  it('should create 100 timedProvider pools', async () => {
+  it('should create 100 simple refund pools with timedProvider', async () => {
     const userCount = '100';
     const params = _createProviderParams(timedProvider.address);
     await _testMassPoolsData(timedProvider.address, amount, userCount, params);
